improve VGA font display on many sizes by enforcing glyphs to 8x16 grid
Glyphs using the 9th "pixel" column, like the box drawing ones, were edited to
use be only 8 "pixels" wide. `EM` size was also changed from 2048 to 16,
enforcing all glyphs to an 8x16 grid.

As a result, now the font on both xterm and gnome-terminal looks perfect not
only on sizes 12 and 21 but also 18 and 30, and all other sizes multiple of 3
like 15, 21, 27, 33, look great.